move new class to test utils
