+++ flexmeasures/app.py
@@ -20,6 +20,7 @@ from redis import Redis
 from rq import Queue
 
 from flexmeasures.data.services.job_cache import JobCache
+from flexmeasures.tests.utils import RQCompatibleFakeStrictRedis
 
 
 def create(  # noqa C901
@@ -89,15 +90,6 @@ def create(  # noqa C901
 
     # configure Redis (for redis queue)
     if app.testing:
-        from fakeredis import FakeStrictRedis
-
-        class RQCompatibleFakeStrictRedis(FakeStrictRedis):
-            def client_list(self, *args, **kwargs):
-                clients = super().client_list(*args, **kwargs)
-                for client in clients:
-                    client.setdefault("addr", "fakeredis:0")
-                return clients
-
         redis_conn = RQCompatibleFakeStrictRedis(
             host="redis", port="1234"
         )  # dummy connection details
